Issue: #19 Technical Details: - Refactored code into base pipeline Blocks - Corrected pipeline configuration in ConfigureSitecore.cs
